backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
requests
pydantic
pdfplumber
PyPDF2
python-docx
python-pptx
pdf2image
pytesseract
Pillow
tiktoken
